dalsia praca na api a komm. medzi servicemi

+++ bet-service/src/main/java/com/example/service/BetslipService.java
@@ -1,5 +1,7 @@
 package com.example.service;
 
+import com.example.commons.Guess;
+import com.example.commons.MatchStatus;
 import com.example.entity.Bet;
 import com.example.entity.Betslip;
 import com.example.models.*;
@@ -10,22 +12,25 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class BetslipService {
     private final BetslipRepository betslipRepository;
     private final DraftBetslipService draftBetslipService;
     private static final double DEFAULT_STAKE = 0.5;
-    public BetslipService(BetslipRepository betslipRepository,DraftBetslipService draftBetslipService) {
+    private final MatchStatusFeignClient matchStatusFeignClient;
+    public BetslipService(BetslipRepository betslipRepository,DraftBetslipService draftBetslipService,MatchStatusFeignClient matchStatusFeignClient) {
         this.betslipRepository = betslipRepository;
         this.draftBetslipService = draftBetslipService;
+        this.matchStatusFeignClient = matchStatusFeignClient;
     }
 
     public BetslipDTO getBetslipById(Long id){
         return betslipRepository.findById(id).map(BetslipMapper::maptoDTO).orElseThrow(EntityNotFoundException::new);
     }
 
-    public DraftBetslip createBetslip(Long userId){
+    public CreateBetslipResponse createBetslip(Long userId){
 //        Betslip betslip = new Betslip(userId,betslipRequest.stake());
 //
 //        for (CreateBetRequest br : betslipRequest.betRequests()){
@@ -42,22 +47,52 @@ public class BetslipService {
         for (CreateBetRequest br : draftBetslip.bets()){
             betslip.addBet(new Bet(br.matchId(),br.guess()));
         }
-        betslipRepository.save(betslip);
-        draftBetslipService.deleteDraft(userId);
-        DraftBetslip newDraft = new DraftBetslip(
-                userId,
-                DEFAULT_STAKE,
-                new ArrayList<>()
-        );
 
-        draftBetslipService.saveDraft(userId, newDraft);
-        return newDraft;
-    }
+        List<MatchStatus> matchStatuses = new ArrayList<>();
+        betslip.getBets().forEach(ms-> matchStatuses.add(matchStatusFeignClient.getMatchStatus(ms.getMatchId())));
+        if (matchStatuses.stream().allMatch(m -> MatchStatus.NOTSTARTED == m)) {
+            betslipRepository.save(betslip);
+            draftBetslipService.deleteDraft(userId);
+            DraftBetslip newDraft = new DraftBetslip(
+                    userId,
+                    DEFAULT_STAKE,
+                    new ArrayList<>()
+
+            );
+            draftBetslipService.saveDraft(userId, newDraft);
+            return new CreateBetslipResponse(newDraft,new ArrayList<>());
+
+        }
+        else{
+            List<Long> matchesStarted = new ArrayList<>();
+            List<Bet> filteredBets = new ArrayList<>();
+            betslip.getBets().stream().forEach((b)-> {if (matchStatusFeignClient.getMatchStatus(b.getMatchId()) == MatchStatus.NOTSTARTED)
+                filteredBets.add(b);
+                else{matchesStarted.add(b.getMatchId());}
+            });
+            betslip.setBets(filteredBets);
+            List<CreateBetRequest> createBetRequests = filteredBets.stream().map(b -> new CreateBetRequest(b.getMatchId(),b.getGuess())).toList();
+            DraftBetslip updated = new DraftBetslip(
+                    draftBetslip.id(),
+                    draftBetslip.stake(),
+                    createBetRequests
+            );
+            draftBetslipService.saveDraft(userId,updated);
+            return new CreateBetslipResponse(updated,matchesStarted);
+            }
+        }
+
+
 
     public DraftBetslip addBetToDraft(CreateBetRequest createBetRequest, Long userId) {
 
+        MatchStatus matchStatus = matchStatusFeignClient.getMatchStatus(createBetRequest.matchId());
+        if (matchStatus != MatchStatus.NOTSTARTED)
+            throw new IllegalStateException("This match is not available for betting");
         DraftBetslip draft = getUsersDraftBetslip(userId);
-
+        boolean alreadyExists = draft.bets().stream().anyMatch((br)-> Objects.equals(br.matchId(), createBetRequest.matchId()));
+        if (alreadyExists)
+            return updateBetOnBetslip(userId,createBetRequest.matchId(),createBetRequest.guess());
         List<CreateBetRequest> updatedBets =
                 new ArrayList<>(draft.bets());
 
@@ -113,5 +148,16 @@ public class BetslipService {
         draftBetslipService.saveDraft(userId, newDraft);
 
         return newDraft;
+    }
+
+    public DraftBetslip updateBetOnBetslip(Long userId, Long matchId, Guess guess){
+        DraftBetslip draftBetslip = getUsersDraftBetslip(userId);
+        CreateBetRequest createBetRequest = draftBetslip.bets().stream().filter(cbr -> Objects.equals(cbr.matchId(), matchId)).findFirst().orElse(null);
+        if (createBetRequest != null){
+            draftBetslip.bets().remove(createBetRequest);
+            draftBetslip.bets().add(new CreateBetRequest(matchId,guess));
+        }
+        draftBetslipService.saveDraft(userId,draftBetslip);
+        return  draftBetslip;
     }
 }
